xll：修改实体类字段，添加订单列表分页查询

+++ common-api/src/main/java/com/tj/product/HappysysDeadline.java
@@ -21,6 +21,6 @@ public class HappysysDeadline {
  private Integer deadlineId;
  private Integer deadlineDays;         //真实天数
  private String deadlineTime;          //显示的值
- private Integer deadlinePrice;        //价格
+ private Double deadlinePrice;        //价格
 
 }
